Начал делать GameRepository

+++ Assets/Scripts/Resources/Stone/StoneObserver.cs
@@ -21,7 +21,7 @@ namespace GameEngine
             _disposable.Dispose();
         }
 
-        private void OnStoneChanged(long wood)
+        private void OnStoneChanged(int wood)
         {
             _view.UpdateCurrency(wood);
         }
